requirements.txt
ipsetpy
paho-mqtt
python-iptables
pyptables
schedule
